+++ tests/core/statemachine_test.py
@@ -9,7 +9,7 @@ class PrintState(State):
 
     _text: str
 
-    def __init__(self, name, text):
+    def __init__(self, text, name = ""):
         super().__init__(name)
         self._text = text
 
@@ -24,10 +24,10 @@ class DummyState(State):
 
 
 def test_simple_machine(capsys):
-    ps1 = PrintState("ps1", "print1")
-    ps2 = PrintState("ps2", "print2")
+    ps1 = PrintState("print1")
+    ps2 = PrintState("print2", name="ps2")
     ps1.add_transition_on_success(ps2)
-    exe = Machine("xe", ps1, end_state_ids=["ps2"], rate=10)
+    exe = Machine(ps1, end_state_ids=["ps2"], rate=10)
     b = Board()
     exe.run(b)
     assert exe.is_end()
@@ -40,12 +40,12 @@ def test_transition_on_failed(capsys):
         def execute(self, board):
             return StateStatus.FAILED
 
-    fs = failedState('fs')
-    ps1 = PrintState("ps1", "success")
-    ps2 = PrintState("ps2", "failed")
+    fs = failedState()
+    ps1 = PrintState("success", name="ps1")
+    ps2 = PrintState("failed", name="ps2")
     fs.add_transition_on_success(ps1)
     fs.add_transition_on_failed(ps2)
-    exe = Machine("xe", fs, end_state_ids=["ps1", "ps2"], rate=10)
+    exe = Machine(fs, end_state_ids=["ps1", "ps2"], rate=10)
     exe.run(None)
     assert exe.is_end()
     assert exe._curr_state.check_name("ps2")
@@ -65,7 +65,7 @@ def test_transition_on_complete(capsys):
     ds1.add_transition_on_success(ns)
     ns.add_transition_on_complete(ds2)
 
-    exe = Machine("xe", ds1, end_state_ids=["ds2"], rate=10)
+    exe = Machine(ds1, end_state_ids=["ds2"], rate=10)
     exe.run(None)
     assert exe.is_end()
     assert exe._curr_state.check_name("ds2")
@@ -73,12 +73,12 @@ def test_transition_on_complete(capsys):
 
 
 def test_simple_machine2(capsys):
-    ps1 = PrintState("ps1", "print1")
-    ps2 = PrintState("ps2", "print2")
-    ps3 = PrintState("ps3", "print3")
+    ps1 = PrintState("print1", name="ps1")
+    ps2 = PrintState("print2", name="ps2")
+    ps3 = PrintState("print3", name="ps3")
     ps1.add_transition_on_success(ps2)
     ps2.add_transition_on_success(ps3)
-    exe = Machine("xe", ps1, rate=10)
+    exe = Machine(ps1, rate=10)
     b = Board()
     exe.start(b, manual_exec=True)
     assert capsys.readouterr().out == "print1\n"
@@ -94,7 +94,7 @@ def test_chain_case():
     s1.add_transition_on_success(s2)
     s3 = DummyState("s3")
     s2.add_transition_on_success(s3)
-    exe = Machine("xe", s1, end_state_ids=["s3"], rate=10)
+    exe = Machine(s1, end_state_ids=["s3"], rate=10)
     b = Board()
     exe.start(b, manual_exec=True)
     exe.update(b, True)
@@ -107,7 +107,7 @@ def test_end_case():
     ps1 = PrintState("ps1", "Hello World")
     es = DummyState("endState")
     ps1.add_transition_on_success(es)
-    exe = Machine("xe", ps1, end_state_ids=["endState"], rate=10)
+    exe = Machine(ps1, end_state_ids=["endState"], rate=10)
     exe.run()
     assert exe.is_end()
 
@@ -115,7 +115,7 @@ def test_end_case():
 # This test checks if machine is completed (is_end())
 # only after the endState finish execution.
 def test_end_case_delay(capsys):
-    ps1 = PrintState("ps1", "Hello World")
+    ps1 = PrintState("Hello World")
 
     class EndState(State):
         def execute(self, board):
@@ -124,19 +124,19 @@ def test_end_case_delay(capsys):
             return StateStatus.SUCCESS
     es = EndState('endState')
     ps1.add_transition_on_success(es)
-    exe = Machine("xe", ps1, end_state_ids=["endState"], rate=10)
+    exe = Machine(ps1, end_state_ids=["endState"], rate=10)
     exe.run()
     assert exe.is_end()
     assert capsys.readouterr().out == "Hello World\ncompleted\n"
 
 
 def test_machine_rate_slow():
-    w1 = WaitState("w1", 0.1) # execute at second 0
-    w2 = WaitState("w2", 0.1) # execute at second 2
+    w1 = WaitState(0.1) # execute at second 0
+    w2 = WaitState(0.1) # execute at second 2
     es = DummyState("endState")  # execute at second 4
     w1.add_transition_on_success(w2)
     w2.add_transition_on_success(es)
-    exe = Machine("xe", w1, end_state_ids=["endState"], rate=0.5)
+    exe = Machine(w1, end_state_ids=["endState"], rate=0.5)
     start_time = time.time()
     exe.run()
     duration = time.time() - start_time
@@ -146,12 +146,12 @@ def test_machine_rate_slow():
 
 
 def test_machine_rate_fast():
-    w1 = WaitState("w1", 0.05) # execute at second 0
-    w2 = WaitState("w2", 0.05) # execute at second 0.1s
+    w1 = WaitState(0.05) # execute at second 0
+    w2 = WaitState(0.05) # execute at second 0.1s
     es = DummyState("endState")  # execute at second 0.2
     w1.add_transition_on_success(w2)
     w2.add_transition_on_success(es)
-    exe = Machine("xe", w1, end_state_ids=["endState"], rate=10)
+    exe = Machine(w1, end_state_ids=["endState"], rate=10)
     start_time = time.time()
     exe.run()
     duration = time.time() - start_time
@@ -161,21 +161,21 @@ def test_machine_rate_fast():
 
 def test_nested_machine(capsys):
 
-    ps1 = PrintState("ps1", "in mach1")
-    ps15 = PrintState("ps15", "in mach1 too")
+    ps1 = PrintState("in mach1")
+    ps15 = PrintState("in mach1 too")
     es = DummyState("endState")
     ps1.add_transition_on_success(ps15)
     ps15.add_transition_on_success(es)
-    mach1 = Machine("xe", ps1, end_state_ids=["endState"], rate=10)
+    mach1 = Machine(ps1, end_state_ids=["endState"], rate=10)
 
-    ps2 = PrintState("ps2", "enter mach1")
+    ps2 = PrintState("enter mach1")
     ps2.add_transition_on_success(mach1)
-    ps3 = PrintState("ps3", "leaving mach1")
+    ps3 = PrintState("leaving mach1")
     mach1.add_transition_on_success(ps3)
     es2 = DummyState("endState")
     ps3.add_transition_on_success(es2)
 
-    mac2 = Machine("mac2", ps2, end_state_ids=["endState"], rate=10)
+    mac2 = Machine(ps2, end_state_ids=["endState"], rate=10)
     mac2.run()
 
     assert capsys.readouterr().out == "enter mach1\nin mach1\nin mach1 too\nleaving mach1\n"
@@ -189,14 +189,14 @@ class RaiseExceptionState(State):
 
 def test_machine_with_exception(capsys):
 
-    ps1 = PrintState("ps1", "p1")
+    ps1 = PrintState("p1")
     re1 = RaiseExceptionState('re1')
-    ps2 = PrintState("ps2", "p2")
+    ps2 = PrintState("p2")
 
     ps1.add_transition_on_success(re1)
     re1.add_transition_on_success(ps2)
 
-    mac = Machine("mac", ps1, ["ps2"])
+    mac = Machine(ps1, name="mac", end_state_ids=["ps2"])
     mac.run()
 
     assert capsys.readouterr().out == 'p1\n'
@@ -212,7 +212,7 @@ def test_machine_with_exception_in_transition(capsys):
 
     is1.add_transition(lambda s, b: s.unknown(), is2)
 
-    mac = Machine("mac", is1, ["is2"])
+    mac = Machine(is1, ["is2"])
     mac.run()
 
     assert mac._status == StateStatus.EXCEPTION
@@ -223,12 +223,12 @@ def test_machine_with_exception_in_transition(capsys):
 
 def test_machine_with_exception_in_transition_with_zombie_states(capsys):
 
-    ws1 = WaitState('ws1', 10)
+    ws1 = WaitState(10)
     is2 = DummyState('d2')
 
     ws1.add_transition(lambda s, b: s.unknown(), is2)
 
-    mac = Machine("mac", ws1, ["is2"])
+    mac = Machine(ws1, ["is2"])
     mac.run()
     assert mac._status == StateStatus.EXCEPTION
     # this is an interrupted, because exception happen at higher level
@@ -245,10 +245,10 @@ def test_debugging_machine(caplog):
 
     logger = logging.getLogger(__name__)
 
-    s1 = WaitState('s1', 1.1)
+    s1 = WaitState(1.1, name='s1')
     s2 = DummyState('s2')
     s1.add_transition_on_success(s2)
-    mac = Machine("mac", s1, ["s2"], debug=True, rate=1, logger=logger)
+    mac = Machine(s1, name="mac", end_state_ids=["s2"], debug=True, rate=1, logger=logger)
     mac.run()
     assert mac.is_end()
     assert len(caplog.records) == 3
@@ -258,10 +258,10 @@ def test_debugging_machine(caplog):
 
 
 def test_interrupt_machine(capsys):
-    s1 = WaitState('s1', 1.1)
+    s1 = WaitState(1.1)
     s2 = DummyState('s2')
     s1.add_transition_on_success(s2)
-    mac = Machine("mac", s1, ["s2"], debug=True, rate=1)
+    mac = Machine(s1, ["s2"], debug=True, rate=1)
     mac.start(None)
     assert mac.interrupt()
 
@@ -275,8 +275,8 @@ def test_flow_into_machine(capsys):
             print("only-state")
             return StateStatus.SUCCESS
 
-    os = OnlyState("only")
-    mac = Machine("mac", os, ["only"])
+    os = OnlyState(name="only")
+    mac = Machine(os, end_state_ids=["only"])
     mac.run(flow_in=test_phrase)
     assert mac.is_end()
     assert capsys.readouterr().out == "only-state\n"
@@ -305,7 +305,7 @@ def test_repeat_node_in_machine():
     ds4.add_transition_on_success(ds5)
     ds5.add_transition_on_success(ds1)
 
-    exe = Machine('exe', ds1, rate=5)
+    exe = Machine(ds1, rate=5)
     exe.start(None)
     for i in range(1, 6):
         time.sleep(1)